backend/requirements/base.txt
Flask==0.11.1
Flask-Mail==0.9.1
Flask-Migrate==1.7.0
Flask-SQLAlchemy==2.1
Flask-Script==2.0.5
bcrypt==3.1.4
psycopg2==2.7.6
Jinja2==2.8
pyjwt==1.4.0
marshmallow==2.15.1
flask-cors==2.1.2
flask-bcrypt==0.7.1
boto3==1.2.3
raven[flask]==6.9.0
stripe==1.31.1
flask-admin==1.4.0
flask-login==0.3.2
Pillow==2.9.0
slackclient==1.0.0
mailjet_rest==1.1.1
pyquery==1.2.13
jsonschema==2.5.1
python3-keyczar==0.71rc0
requests==2.20.0

elasticsearch==6.3.1
elasticsearch-dsl==6.3.1

chargebee==2.1.2
gunicorn==19.5.0

geoip2==2.4.0

mixpanel==4.3.2

retrying==1.3.3
